split link overlay into two components, fix bug
