+++ src/serlo-editor-repo/plugin-text/components/link/utils.ts
@@ -1,0 +1,3 @@
+export function checkSerloIdHref(input: string) {
+  return input.startsWith('/') && input.slice(1).match(/^\d+$/)
+}
